+++ R/subtype.model.R
@@ -1,4 +1,0 @@
-#' Model for predicting methylation subtype
-#'
-#' [pamr](https://cran.r-project.org/web/packages/pamr/index.html) model used for assigning new patients to four prostate cancer DNA methylation subtypes.
-'subtype.model'