+++ R/validate.subtype.model.cpgs.R
@@ -1,9 +1,9 @@
 #' Validate input data for estimate.subtypes()
 #'
-#' Check whether `methy.data` contains all CpGs required by \link{subtype.model} for assigning patients to four prostate cancer DNA methylation subtypes.
+#' Check whether `methy.data` contains all CpGs required by \link{subtype.model.pamr} or \link{subtype.model.rf} for assigning patients to four prostate cancer DNA methylation subtypes.
 #'
 #' @param methy.data A data.frame with patients as rows (rownames give patient ids) and column names give CpG ids.
-#' @param prop.missing.cutoff The maximum proportion of missing values allowed for each required CpG. KNN imputation is used to impute missing values.
+#' @param prop.missing.cutoff The maximum proportion of missing values allowed for each required CpG.
 #' @export
 #' @return
 #' * `val.passed` a logical indicating whether the data passed validation
@@ -29,8 +29,8 @@ validate.subtype.model.cpgs <- function(methy.data, prop.missing.cutoff = 0.3) {
     methy.data.nomiss <- na.omit(methy.data);
     stopifnot('All values of methy.data should be between 0 and 1' = all(methy.data.nomiss >= 0 & methy.data.nomiss <= 1));
 
-    data(subtype.model, envir = environment());
-    required.cpgs <- rownames(subtype.model$centroids);
+    data(subtype.model.pamr, envir = environment());
+    required.cpgs <- rownames(subtype.model.pamr$centroids);
     missing.cpgs <- setdiff(required.cpgs, colnames(methy.data));
     nonmissing.cpgs <- setdiff(required.cpgs, missing.cpgs);
     if (length(nonmissing.cpgs) > 0) {
@@ -47,10 +47,17 @@ validate.subtype.model.cpgs <- function(methy.data, prop.missing.cutoff = 0.3) {
                 }
             }
         );
+
+    # regardless of what the user specifies for prop.missing.cutoff, we should
+    # print a warning if some CpGs have high missing.
+    cpgs.high.miss.warn <- sum(required.cpgs.prop.missing > 0.5);
+    if (cpgs.high.miss.warn > 0) {
+        message('Warning: ', cpgs.high.miss.warn, ' out of ', length(required.cpgs) ,' required CpGs have > 50% missing values. Having many CpGs with high missing data may decrease accuracy of subtype assignment.');
+        }
+
     required.cpgs.with.high.missing <- lapply(required.cpgs.prop.missing, function(x) x[x > prop.missing.cutoff]);
     val.passed <- length(unlist(required.cpgs.with.high.missing)) == 0 & length(unlist(missing.cpgs)) == 0;
-    val.passed;
-    unlist(required.cpgs.with.high.missing)
+
     return(list(
         val.passed = val.passed,
         required.cpgs = required.cpgs,