+++ inst/subtype.model.R
@@ -1,8 +1,19 @@
+# there are 2 models for assigning the 4 methylation subtypes to new samples: PAMR and random forest
+
 source('config.R') # see project PRAD-000101-MethySubtypes/PrCaMethy/config.R
-load(arg$path.subtype.model);
 
-subtype.model <- model.for.predicting.subtypes;
-subtype.model.required.cpgs <- rownames(subtype.model$centroids);
 
-usethis::use_data(subtype.model, overwrite = TRUE, compress = 'xz');
+### PAMR
+load(arg$path.subtype.model.pamr);
+
+subtype.model.pamr <- model.for.predicting.subtypes;
+required.cpgs <- rownames(subtype.model.pamr$centroids);
+
+usethis::use_data(subtype.model.pamr, overwrite = TRUE, compress = 'xz');
 usethis::use_package('pamr');
+
+### Random Forest
+load(arg$path.subtype.model.rf);
+subtype.model.rf <- rf;
+usethis::use_data(subtype.model.rf, overwrite = TRUE, compress = 'xz');
+usethis::use_package('randomForestSRC');